refactor: Extract useVideoLoader hook from god hook
- Create useVideoLoader hook (451 lines) for video loading logic
- Extract: loadVideoSafely, loadVideo, handleVideoUpload, loadSampleVideo
- Extract: loadHardcodedVideo, loadPistolSquatSample, resetVideoState
- Reduces useExerciseAnalyzer from 1515 to 1212 lines (20% reduction)
- Hook manages abort controllers and URL cleanup internally
- Maintains same API for consumers
